Change results page to show product thumbnails, no more buylist
The buylist is really only used in very special cases, such as a quick-
order feature. The intent of the results page as used is to offer
browsing through a category or product group.

Also, the buylist had an old poorly-supported "munge" feature, so if a
quick-order page is desired, it should be written from scratch.